Change bill data type to a float number

+++ day2/tip_calculator.py
@@ -11,7 +11,7 @@ tip = input("How much tip would you like to give? 10, 12, or 15 percent?\n%")
 num_people = input("How many people to split the bill?\n")
 
 # Calculate how much each person should pay
-total_bill = int(bill) + (int(tip) / 100) * int(bill) # print(type(total_bill)) <class 'float'>
+total_bill = float(bill) + (int(tip) / 100) * float(bill) # print(type(total_bill)) <class 'float'>
 each_person_bill = total_bill / int(num_people)
 print(f"Each person should pay: ${round(each_person_bill, 2)}")
